+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # ruby_gem_criterion_channel CHANGELOG
 
-## October 16, 2020: Release Tags
+## December 15. 2021: Added Search
+- Added search using Mechanize to return search results of titles
+available in the Criterion Channel catalog
+
+## October 16, 2021: Release Tags
 - Tagged 0.0.1 in preparation for creation of new version of gem to allow for more semantic versioning.
 - Created branch for of 0.1.0 (semantic versioning/allow for bumping up of minor release updates for dependencies).
 - Merge branch of 0.1.0 into maaster branch